MB-10794: Set the open checkpoint seqno properly when backfill ends
When we initially create a replica vbucket we create an open
checkpoint with checkpoint id 0 and this checkpoint will start
at seqno 1. Since this replica might receive backfill items when
data is replicated to it we need to properly reassign the seqno
of the start checkpoint item before once backfill is finished.

Change-Id: Ic04998b5cd2712d786e07922fe82be83d2c8b322
Reviewed-on: http://review.couchbase.org/35486
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>