MB-10794: Set the open checkpoint seqno properly when backfill ends
When we initially create a replica vbucket we create an open
checkpoint with checkpoint id 0 and this checkpoint will start
at seqno 1. Since this replica might receive backfill items when
data is replicated to it we need to properly reassign the seqno
of the start checkpoint item before once backfill is finished.

Change-Id: I9c224f6b0d517fe500b733ce84cc00366b702a2c
Reviewed-on: http://review.couchbase.org/35708
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>