Notify memcached when an upr stream is set to dead state
If a stream is currently not sending any items then setting the
stream to dead state (which adds an end stream message) will not
cause the memcached worker thread to visit the upr stream in order
to send this message. As a result we need to notify memcached when
streams die in order to make sure that consumers don't get stuck
waiting for the end stream message.

Change-Id: Iad46cd18d6d911e8826c0ff235924a8403a90c54
Reviewed-on: http://review.couchbase.org/35748
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>