Fix race conidtion when deleting vbuckets
There is a race condition where we create, delete, and then create
a vbucket. The problem is that the the first create can finish
after the second create is started. When this happens the code
thinks that the last create has finished even though the vbucket
deletion has not. This can cause items to fail to persist to disk.

Change-Id: I66e6450481acda2250bb7e13f921f3e9d20fbee1
Reviewed-on: http://review.couchbase.org/36254
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>