Add items to retry queue when vbucket is in deletion state
It is possible that ep-engine may create, delete, and then recreate
a vbucket in a short timespan during vbucket takeover. When this
happens both the vbucket creation and deletion flag will be set an
this means that it is possible that the last created vbucket will
be visited by the flusher. If this happens the flusher will drop the
items from the vbucket and result in items not getting persisted to
disk. This change causes items that the flusher is trying to persist
to be re-queued for persistence when we are in this state.

Change-Id: I34ea9208abddad34d9506c20ee7e693f957d2de9
Reviewed-on: http://review.couchbase.org/37018
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>