MB-10979: set sequence number properly for temporary initial item
When a temporary initial item is being added to the hash table,
the sequence number needs to be set to the max deleted sequence
number as opposed to being zero. There is a possibility that
a set from the client can access this temporary item before
the BG fetcher task can update the metadata with the correct
sequence number.

Change-Id: Iebf5dc70c2cc25ed4a4cf5f6540f1848772491cf
Reviewed-on: http://review.couchbase.org/37104
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>