MB-11104 Byseq num of checkpoint_end shouldn't be > current_val
The byseq num of checkpoint_end item shouldn't be greater than
the current value. Otherwise, the cursor sitting on checkpoint_end
item in an old checkpoint might miss the first regular item in a
collapsed checkpoint. I confirmed that this won't cause any
issues in registering the cursors based on their seq numbers.

Change-Id: I6a15d0b896b3fe7e6e73979572369b5c8552269e
Reviewed-on: http://review.couchbase.org/37358
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>