MB-11219: Always send a snapshot marker even if the snapshot has no items
It is possible that for a seqno range there are no items. Even if
this happens we still need to send a snapshot marker to signify
what range the engine is trying to send. This is particularly a
problem when we do a vbucket takeover and there are items to send
during the takeover. If the last snapshot sent was a disk snapshot
then the new master vbucket can be left in backfilled state and
this will cause problems for persistence.

Change-Id: I6b5fade5195395a5f7ac0fe722336d44175e5884
Reviewed-on: http://review.couchbase.org/37669
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>