MB-11085: Always create a new failover entry on unclean shutdowns
In the past we wouldn't generate a new failover entry if the high
seqno number on disk was the same after a crash. This is incorrect
because it is possible that the server did receive mutations and
replicated them without persisting them before the crash. If this
happens the consumers of upr streams will no roll back their data
properly because the failover entry will not change on the server.

Change-Id: I8c6bab504f0be3298e1e888dbe6f3fac9c3fa905
Reviewed-on: http://review.couchbase.org/37670
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>