MB-11254 remove unnecessary assert unlockShardTask
This assert fires if unlockShardTask is called when the shard
is already unlocked.
If a shard-serialized writer task such as flusher gets
cancelled even before it got to run, it might end up
calling unlockShardTask() even though it never locked it.
This is ok - since the task is going away, so removing this assert.

Change-Id: Ife2c98611327e1d9d3e926b2252242231b2e48e5
Reviewed-on: http://review.couchbase.org/37689
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>