MB-11234: Notify when item expired by GET
When we try GET-ing an item and find that it has
expired already, we issue a softDelete on it,
upon which the item is added to the dirty queue.
When enqueued, we should notifyVBConnections for
the producer to notify that the item is waiting
to be replicated and is currently in the producer
queue.

Change-Id: I0d83a2d0baabb7a9f992f8feb802cd34633079e9
Reviewed-on: http://review.couchbase.org/38212
Reviewed-by: Sundararaman Sridharan <sundar@couchbase.com>
Tested-by: abhinav dangeti <abhinav@couchbase.com>