MB-11361: Implement upr noop command
The upr noop command is used to detect dead connections. It is sent
once every 3 minutes and waits 3 minutes for a response. If no
response comes in that time the connection is deemed to be dead. The
noop command is only sent when flow control is turned on by the
consumer.

Change-Id: I2eb9fc201eeb91878f9339c6b3b582cd2ca5da06
Reviewed-on: http://review.couchbase.org/38439
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>