MB-11361: Implement upr noop command
The upr noop command is used to detect dead connections. It is sent
once every 3 minutes and waits 3 minutes for a response. If no
response comes in that time the connection is deemed to be dead. The
noop command is only sent when flow control is turned on by the
consumer.

Change-Id: Ib6f1eb51810d9c641ca52d1e3cda8ae9ffa3684c
Reviewed-on: http://review.couchbase.org/38355
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>