MB-11067: Reduce initial IOV list size to 10
Measuring the high watermark of how many iov elements are used for
sendmsg(), we find that during beer-sample load and pillowfight tests,
the highest used number of IOVs is only 5 (2-node cluster, replication
enabled).  Running a TAP backup this increases to 20.

Therefore (significantly) reduce the current initial size of 400 down
to 10 - normal KV connections are unlikely to need to resize any
larger, and TAP connections will only need to resize once.

Similary, reduce the high watermark for buffer shrinking from 600 down
to 50.

This brings the size of a connection object down to 1552 bytes.

Change-Id: Id6d2d96cc94f2ec64ab8f9083abc033dcaa4d82a
Reviewed-on: http://review.couchbase.org/38671
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>