MB-11067: Reduce initial msghdr list size to 5
Measuring the high watermark of how many msglist elements are used for
sendmsg(), we find that during beer-sample load and pillowfight tests,
the highest used number of IOVs is only 4 (2-node cluster, replication
enabled). This stays the highest for TAP and UPR backup, and for rebalance.

Therefore reduce the current initial size of 10 down to 5 - From what
I can observe this should be sufficient for most connections, not
requiring any expansion (which is handled if necessary by
add_msghdr().

Similary, reduce the high watermark for buffer shrinking from 100 down
to 20.

This brings the size of a connection object down to 1312 bytes.

Change-Id: I38d10dad8c46a6880b60acabe6b654a40fdcb756
Reviewed-on: http://review.couchbase.org/38680
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>