MB-11085: Always generate a failover entry after an unclean shutdown
Previously we would not generate a new failover entry after an
unclean shutdown if the persisted seqno did not change between
when the server was originally started and when is crashed. This can
cause issues because it is possible that mutations were received in
memory and streamed to another node and then the server crashes. If
we do not generate a new failover entry in this case then we will
not properly tell the node that received the mutations from memory
to rollback.

Change-Id: I6a97c12f2eb300984aaa81d7f9fdd0eb3272233d
Reviewed-on: http://review.couchbase.org/38698
Reviewed-by: David Liao <david.liao@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>