Initialize control cas to 0xdeadbeef
If its default value is 0 there is no way to change it without
using 0 as the cas value (which means override) introducing
the possibility for race conditions. With a non-null value it
may be changed by someone who requested its value and protects
against race conditions.

Allow for using cas value 0 to override cas for setting the cas
(to make it behave consistently with the rest of the cas usage)

Change-Id: I637104df4f65527d8ad10636c01a0e7f53930c53
Reviewed-on: http://review.couchbase.org/38784
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>