MB-11579: Use the upr latest flag to set the end seqno to the vb high seqno
The "latest" flag is used to create a stream and get everything up to the
latest mutation that the vbucket currently contains. This flag should be
used by clients sho are interested in getting the latest mutations, but
do not want to have to figure out what the latest seqno is by using stats.

Change-Id: I86b6b3cebab46351b41ecdc11194c5b5607a4f28
Reviewed-on: http://review.couchbase.org/38937
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>