MB-11387: Handle backfill condition where end_seqno_ is equal to curChkSeqno
There is a possibility that the when the end_seqno_ is the same as the current
checkpoint seqno, the flusher has only persisted items till end_seqno_ - 1.
Given that the item that has curChkSeqno will be streamed from memory anyways,
the backfill can just be scheduled to curChkSeqno - 1. This will ensure that
the backfill task is not blocked until the item with end_seqno_ is flushed to disk.

Change-Id: I97c957967f8914403647ce0185ff0f0c2bab66e4
Reviewed-on: http://review.couchbase.org/39002
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>