MB-11583 avoid cleanup race on force disconnect
When there is a stream created with same name, then we do a
disconnect and resulting in immediate cleanup for that connection.
If this cleanup thread races with the connection's step routine
it may read an invalidated cookie causing an invalid pointer access
on the step() routine seen in the crash.
Fix is to only set the connection state to disconnected and have the
callback do the actual cleanup of the connection cookies.
Change-Id: I12b2ed93b9f8bd04cbc79c953bacd76171b8745f
Reviewed-on: http://review.couchbase.org/39027
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>