MB-11621: Grab streamMutex before reconnecting stream during rollback
When we reconnect a stream we notify the memcached layer that the
stream has been reconnected. The notification mecachism in the
consumer assumes that we already are holding the stream lock so we
need to make sure we grab it in order to avoid a race condition we
modifying the ready queue.

Also note that the opaqueMap relies on the streamMutex which is why
we grab it where we do.

Change-Id: Ia65c0ab2db3500f7480818f849000bc5ee3db0af
Reviewed-on: http://review.couchbase.org/39181
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>