MB-11658 compactor must purge tombstones if drop_deletes set
This must be done unconditionally. Current logic will not
purge any tombstones if ns_server does not set the drop_delete flag.
However in 2.5 the expectations is that if
drop_delete flag in compactor is set, delete all tombstones unconditionally
if drop_delete flag is not set, purge tombstones based on purge_seqno value
Change-Id: I5d6060fb3e4c65a7f6850f30db0cf61a9e1c3d92
Reviewed-on: http://review.couchbase.org/39183
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>