MB-11573 Don't check the cached number of items in disk backfill
We should avoid checking the cached number of items from the
read-only kvstore instance because it is not always consistent
with the one on disk. In addition, there is a case where there
are only deleted items in the Couchstore file, which causes the
cached item counter to zero. In this case, we still need to read
deleted items from disk for UPR streaming.

Change-Id: I04c98414ccedb707dd171b050252f1c465c7795d
Reviewed-on: http://review.couchbase.org/39224
Reviewed-by: Sundararaman Sridharan <sundar@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>