MB-11676: Properly set the end seqno for disk only streams
It's possible that if there were no persisted disk items that we
would not schedule a backfill and go straight to the in-memory
phase which is incorrect for a disk only stream. This change also
makes sure that the end seqno is reported correctly in the stats
for a disk only stream.

Change-Id: I30c473149fb37f38e47f5c8c87c7d7db60608223
Reviewed-on: http://review.couchbase.org/39242
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>