MB-10962 Don't snooze the connection notifier unnecessarily.
There is a race in pausing and resuming the connection
notifier upon receiving a new mutation. To address this issue,
This change adds a flag that indicates if there is a pending
notification event for any TAP connections.

Note that this change was originally merged in 2.5.0 branch to
address the sporadic high latency issue in TAP-based replication.

We still need to address it in the UPR-based replication
separately.

Change-Id: Ia8d39220c46647a21254c63f37e7ce3ceaa2bc99
Reviewed-on: http://review.couchbase.org/39250
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>