MB-11553 Adjust the checkpoint creation interval.
This change reduces the checkpoint creation interval to
mitigate the memory overhead caused by having blob instances
inside the open checkpoint. If a checkpoint creation interval
is large, it can cause significant memory overhead due to blob
instances inside the current open checkpoint.

Note that we also limit the number of checkpoints in memory, so
this change won't allow us to keep creating a new checkpoint if
the limit is reached.

Change-Id: I33119ef8b539d5f0209b64bcfcfcbd82632ae7e1
Reviewed-on: http://review.couchbase.org/39333
Reviewed-by: Sundararaman Sridharan <sundar@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>