MB-11661 Disable checkpoint merge support by default.
Merging closed checkpoints in a replica vbucket was mainly
supported to reduce the memory overhead in the replica vbucket,
which performs the replication to another replica vbucket.
For example, in the replication chain A->B->C, merging closed
checkpoints can be performed on the replica B if the replica C
is overloaded and constantly backs off the replication from the
replica B.

As of 3.0, we switch to the star replication, which doesn't
require supporting checkpoint merge operation. Thus, this change
disables the checkpoint merge operation by default, but still
allows us to enable / disable it dynamically at runtime.

Change-Id: I46b17030f35a5dad9575a6ecd2fde33f70b497bc
Reviewed-on: http://review.couchbase.org/39342
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>