MB-11725: Add log message to help find cause of dataloss
There are two potential cases that might be the root cause of data
loss. One is that the parameters supplied to the backfill task are
incorrect and the other is that the backfill is actually done
improperly somewhere in couchstore. This log message will help us
narrow down where the issue is.

Change-Id: I31fd8c001ac220287fb2072b94aef95bede77da1
Reviewed-on: http://review.couchbase.org/39642
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>