MB-11799 Throttle vbucket compaction with the disk write queue size.
If a disk write queue size is above the threshold, it is better to
throttle vbucket compaction tasks, so that the minimum number of
compaction tasks can be running at the same time.

As another commit, we will make both the write queue size threshold and
the minimum number of compaction tasks configurable at runtime.

Change-Id: I70d6c3f02507895ed767060870e9e16265cd60ac
Reviewed-on: http://review.couchbase.org/40043
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>