MB-12179: Remove EPStats from CouchKVStore
We no longer need this dependency in the KVStore and I moved the
only stat that we updated into the KVStore since I think it makes
more sense to have it there anyways. I'm planning on modularizing
the KVStore soon so that we can write unit tests, but am adding
this change now since it simplifies some of the refactoring I did
in following changes.

Change-Id: Ifdd48857e076b57606a65a49521d9021556dd8ac
Reviewed-on: http://review.couchbase.org/41781
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>