Fixup config_parse_test
The test failed on Linux due to the fact that the test
required TMP, TEMP or TMPDIR to be set in environment.
We can always use /tmp or /var/tmp if these directories
don't exist.

In addition the tests didn't flush the output buffer making
the "progress" invisible (and missing a newline at the end)

Change-Id: I3a7c463a38da4dde32472d64fb8aa70c0aa9bba1
Reviewed-on: http://review.couchbase.org/41013
Tested-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>