Extend the search for dynamic engine creation
I'd like to avoid dynamically loading default_engine.so and
ep.so and rather call the functions to create them explicitly.

In order to avoid multiple symbols with the same name I'd like
to rename the creation function.

Moving forward we'll be dropping bucket_engine so we'll be able
to remove the duplication of the code..

Change-Id: I7b329987913a94c5d7c869b5c8e6d0ae2faa4aa4
Reviewed-on: http://review.couchbase.org/41184
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>