Extend the RBAC skeleton with user, roles and profiles
This patch adds basic support for users, roles and profiles.
We haven't fully figured out how we want all of this to work,
so this patch just moves the project in the right direction.
It is still a bare bone system that isn't intended to be used
in a dynamic fashion by the system yet. We're only using
two different settings, the system user with all access and
the "normal" user with access to the most common opcodes (and
no admin rights)

Change-Id: I476a1ac6cfcd9f39ee4b113b87faa6e8fdf0eb06
Reviewed-on: http://review.couchbase.org/41279
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>