MB-12067 [4/6]: Add age field to Blob; use as part of defragmenter policy
When a Blob object is allocated record the current 'age' of it. The
age is incremented when the runs, allowing us to measure how many
defragmenter passes have occurred, and consequently can be used by the
defragmenter to select Blob objects which have been in existance (and
haven't changed size) for a long period, and hence are candiates for
defragmentation.

Change-Id: I6380e6601ae5ee86ca630031e4edea9711a88433
Reviewed-on: http://review.couchbase.org/41497
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>