create syslog compatible logging events
This first commit inserts syslog compatible events.  It uses the
existing interface (so call logging has not changed).  The log
message gets converted into a syslog event.  This event is
then "sent" to the backend logging function, which takes JUST the
event and uses it to generate the same logging messages as
before.

Change-Id: I7a67e35f5111c406ed5c016e11afac2893374169
Reviewed-on: http://review.couchbase.org/41416
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>