Revert "Add alloc_hooks API to mock server"
Reverting as it breaks a number of ep_engine unit tests which rely on
the old "manual" mem_used calculations - by adding the alloc_hooks to
the mock server, the memory tracker is enabled and starts calculating
mem_used from the underlying malloc library.

Intend to re-apply this once the affected testcases are fixed.

This reverts commit 68ceefc431e24475f5272f61051d62091d8291f9.

Change-Id: I4151d6069acdaf1fb8ad397712f8b0fa4811aa90
Reviewed-on: http://review.couchbase.org/41647
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>