Update test_memory_limit to support memory_tracker
When the memory tracker is enabled, mem_used accurately reflects the
actual amount of memory used by the bucket. This test assumes that
mem_used is only the the actual document key & value size, and so when
the memory_tracker is enabled the mem_used figure (correctly)
increases significantly.

Update the testcase to work with much larger initial mem_used, and
calculate the value length based on a fraction of the available space,
not on a fixed overhead.

Change-Id: I95df3544d1569f3e7d1a4efbe18e1f2c9cd99830
Reviewed-on: http://review.couchbase.org/42237
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Tested-by: abhinav dangeti <abhinav@couchbase.com>