Update ep_testsuite mem_used checks to support memory tracker
A number of ep_testsuite tests check for mem_used to *exactly* match
it's value before the test began. This is fragile, particulary when
using the memory tracker as any number of new/delete calls may be
made _in addition_ to the expected stored documents, and hence it is
quite possible the final mem_used value isn't exactly the same.

Update the tests to either check for a relative change, for example
less than the high point, to handle such fluctuations; or remove the
check where the low-level mem_used isn't actually useful in the test.

Change-Id: Ief4d7add1117a4c4eb34d2faf86f7e781c8e5459
Reviewed-on: http://review.couchbase.org/42238
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Tested-by: abhinav dangeti <abhinav@couchbase.com>