Update test_item_pager to support memory tracker
test_item_pager has fixed values for how many documents to
create. This fails when the minimum bucket quota (max_size) must be
increased in-line with the actual amount of memory a bucket takes -
when the memory tracker is enabled.

Update the test to create as many document as possible before TEMP_OOM
is encountered. Also reduce the amount of memory used by temporary
test objects to leave more for actual stored documents.

Change-Id: Id539c7654f81f3cab3a08da6b96521ad388ead58
Reviewed-on: http://review.couchbase.org/42239
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Tested-by: abhinav dangeti <abhinav@couchbase.com>