MB-11373: In case of incorrect datatype set by client
Say the client were to set a binary document or a JSON
document and were to incorrectly set the datatype to a
compressed type, the response handler would return an
error trying to snappy_uncompress an uncompressed document.

In such a scenario, log a warning that says- the
particular key may have an incorrect datatype.

Change-Id: I5e09c5299bda4e3bd094270454b9a2a63040f903
Reviewed-on: http://review.couchbase.org/41263
Reviewed-by: Sundararaman Sridharan <sundar@couchbase.com>
Tested-by: abhinav dangeti <abhinav@couchbase.com>