MB-12067 [jemalloc/Linux]: Implement mem_used tracking using wrapper functions
To be able to track mem_used we need to invoke callbacks whenever
memory is new'd or delete'd. The current allocator (TCMalloc) provides
hooks itself, but jemalloc does not, therefore (for Linux) we create
our own wrapper functions for malloc/calloc/realloc/free and get them
to invoke the hooks, as well as calling into jemalloc to perform the
actual memory (de)allocation.

Change-Id: I3b055f84be569fff43b81877b9ad18cbc67dfb2b
Reviewed-on: http://review.couchbase.org/42522
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>