MB-12179: Pause backfills when after 'x' of items or bytes are read
All dcp backfills are will be paused after reading 4MB or 4096 items
and if any other backfills are scheduled then those will be run before
the original backfill is run again. This change allows us to interleave
backfills so that we no longer have to run a backfill for a single
vbucket to completion before starting the next one.

Change-Id: Ibdb455a5bd931e0276dad61d686f9dbf507ed7af
Reviewed-on: http://review.couchbase.org/42316
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>