Remove unnecessary Item() constuctor #2
Remove another constructor of Item, the one which takes only a data
size - i.e. allocates a value but doesn't copy it in. In all existing
use-cases this can be replaced with either:

a) The constructor which does copy the data in, as the use-case was
   doing performing the copy in two stages (construct Item of correct
   size, then memcpy()).

b) The consturctor which takes both ptr and size, with with ptr==NULL
   - this is already handled by the setData() private method of Item.

Change-Id: Iffbfef423e82bb021f74860a72dc7b5936bf2215
Reviewed-on: http://review.couchbase.org/42625
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Tested-by: abhinav dangeti <abhinav@couchbase.com>