MB-12540: Refactor the way we track snapshot start/end seqnos
For the 3.0 release I needed to add snapshot sequence number
quickly and in doing so implemented an approach that was not
very good and prone to errors. The correct approach is to track
sequence numbers on disk and also to track them for each checkpoint
in the checkpoint manager. This change moves us to the better
approach.

Change-Id: Ifbdd6c87b06767ff8c6504c1058ce11195fa2428
Reviewed-on: http://review.couchbase.org/42861
Reviewed-by: Sundararaman Sridharan <sundar@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>