MB-12745: Disable jemalloc symbol renaming
Depending on how jemalloc was configured for the platform, its header
file may attempt to automatically rename malloc/calloc/free and
friends to je_malloc,je_calloc, etc. This is designed to make it
'easier' for applications to use jemalloc, but in the case of
memcached we define our own versions of malloc/calloc... which
explicitly *do* want to call je_malloc/je_calloc... This renaming
behaviour results in our own interposed functions recusrively calling
themselves and soon crashing.

Fix this by setting JEMALLOC_NO_RENAME to disable this renaming
behaviour; preventing the loop.

Note we still require a jemalloc build which has the je_ prefix, but
this change will at least ensure we have the correct symbol names in
memcached, and prevent us from linking to a library which doesn't have
the correct symbol prefix.

Change-Id: Iecbb0724b3392f4be28585b001cb406f78e7a16e
Reviewed-on: http://review.couchbase.org/43558
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>