MB-12762: Close SSL connection on error during handshake
If a receive error occured during SSL handshake (e.g. received
incorrect handshake data) then instead of closing the connection (as
expected) memcached would belive it needed to retry and return the
connection to the conn_read state. This resulted in the state machine
for the connection essentially looping forever attempting to read
data. This has two observabl side-effects:

1) A connection slot is essentially consumed (forever) by the failed
   connection attempt.

2) The frontend thread that the failed connection was assigned to will
   spin (at 100% CPU), if no other connections owned by the worker have
   any events waiting.

(Note for (2), the CPU isn't "consumed" - other connections if present
*will* be serviced as normal, the issue is simply that the failed
connection immediately posts an event to retry the read, which
libevent will immediately schedule if there are no other events
(i.e. other connections) ready).

Fix is to ensure that the ssl.error status is checked after draining
the SSL recv pipe, and if set then reseting the connection.

Change-Id: I1775b7ad16c6e12fc619da2ba15c93f6e07cd096
Reviewed-on: http://review.couchbase.org/43608
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>