MB-12864: Add Breakpad crash reporting
Add Google's Breakpad project to memcached so fatal errors will
be easier to diagnose, with finer details of system status
nearer to the time of the fatal error than is currently accessible.

- Support for Linux platforms and Windows.

- Mark Woosey's <mark.woosey@couchbase.com> Linux code is
  included, with minor alterations.

- Feature flags ensure memcached can still build without breakpad.

Google Breakpad available at https://code.google.com/p/google-breakpad/

N.B. Breakpad's libraries aren't currently committed and will
require building (and are to be added at a later point).

(This is a re-submit of commit 969ead which caused a build failure
that has been resolved with the following tlm change:
http://review.couchbase.org/#/c/44475/ )

Change-Id: Icba93b8ef7beeeccdb66ce81218cbb1338177bb7
Reviewed-on: http://review.couchbase.org/44476
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>